Improved Requires and BuildRequires parsing
Added the Requirement class to handle Requires and BuildRequires lines
with additional version number specifications. This is intended to
improve upon the current situation where the entire string is a part of
the returned string.

The new class' `name` attribute contains the name of the package which is
the target of the Requires or BuildRequires clause.